Expose licence detection results to the API
So clients can benefit from a rich data structure that contains results
for one or more project paths and that can be parsed as JSON.

The json data structure was already available in the main.go,
so I merely moved it to a top level go file to make it easier to expose.
